Fix a race condition between remove_from_env and other io::process tests.
